Make tests for halls/payments/projections services

+++ Services/THECinema.Services.Data/HallsService.cs
@@ -49,6 +49,12 @@
         public async Task DeleteAsync(int id)
         {
             var hall = this.hallsRepository.All().Where(h => h.Id == id).FirstOrDefault();
+
+            if (hall == null)
+            {
+                throw new ArgumentNullException("The Hall doesn't exist!");
+            }
+
             this.hallsRepository.Delete(hall);
             await this.hallsRepository.SaveChangesAsync();
         }
